Dedupe repo-info and review-response parsing across issue scripts
b_create_issue.py duplicated get_repo_info (already in lib/github_repo.py)
and parse_review_response was byte-for-byte identical to the copy in
n_review_issue.py. Extract the shared parser into lib/issue_review.py and
have both scripts import it and github_repo.get_repo_info instead.

Closes #5759
